refactor: moving state machine hierarchy builder
