+++ src/main/kotlin/at/ac/uibk/dps/cirrina/Runtime.kt
@@ -2,16 +2,15 @@ package at.ac.uibk.dps.cirrina
 
 import at.ac.uibk.dps.cirrina.cirrina.di.CsmMain
 import at.ac.uibk.dps.cirrina.execution.`object`.Context
-import at.ac.uibk.dps.cirrina.execution.`object`.ContextVariable
 import at.ac.uibk.dps.cirrina.execution.`object`.EventHandler
 import at.ac.uibk.dps.cirrina.execution.`object`.Extent
 import at.ac.uibk.dps.cirrina.execution.`object`.StateMachine
+import at.ac.uibk.dps.cirrina.execution.`object`.createHierarchy
 import at.ac.uibk.dps.cirrina.execution.service.RandomServiceImplementationSelector
 import at.ac.uibk.dps.cirrina.execution.service.ServiceImplementation
 import at.ac.uibk.dps.cirrina.execution.service.ServiceImplementationSelector
 import at.ac.uibk.dps.cirrina.io.CsmParser
 import at.ac.uibk.dps.cirrina.spec.Csml as CsmlSpec
-import at.ac.uibk.dps.cirrina.spec.StateMachine as StateMachineSpec
 import io.micrometer.core.instrument.MeterRegistry
 import io.micrometer.core.instrument.Timer
 import jakarta.inject.Inject
@@ -32,7 +31,7 @@ constructor(
   private val stateMachineFactory: StateMachine.Factory,
   persistentContext: Context?,
   meterRegistry: MeterRegistry,
-  @CsmMain csmMainUri: URI,
+  @CsmMain csmMain: URI,
 ) {
   val instances: Map<String, StateMachine>
   val selector: ServiceImplementationSelector
@@ -45,7 +44,7 @@ constructor(
 
   init {
     val spec =
-      CsmlSpec.create(CsmParser.parseCsml(csmMainUri))
+      CsmlSpec.create(CsmParser.parseCsml(csmMain))
         .onFailure { logger.error(it) { "failed to initialize collaborative state machine class" } }
         .getOrThrow()
 
@@ -63,10 +62,12 @@ constructor(
     instances =
       spec.instances
         .flatMap { (name, `class`) ->
-          buildInstances(
+          stateMachineFactory.createHierarchy(
             name,
             spec.collaborativeStateMachine.stateMachines[`class`]
               ?: error("state machine class '${`class`}' not found"),
+            this,
+            selector,
             null,
             spec.subscriptions[name],
             spec.datas[name],
@@ -97,31 +98,4 @@ constructor(
         logger.info { "runtime terminated in $duration" }
       }
   }
-
-  // TODO: Move out of here
-  private fun buildInstances(
-    name: String,
-    specification: StateMachineSpec,
-    parent: StateMachine?,
-    subscriptions: List<String>?,
-    data: List<ContextVariable>?,
-  ): List<StateMachine> =
-    stateMachineFactory
-      .create(name, specification, this, selector, parent, subscriptions, data)
-      .let { current ->
-        specification.nestedStateMachines
-          .flatMapIndexed { index, nestedClass ->
-            buildInstances(
-              "${current.name}.$index@${nestedClass.name}",
-              nestedClass,
-              current,
-              null,
-              null,
-            )
-          }
-          .let { nestedInstances ->
-            current.apply { nested = nestedInstances.map { it.name } }
-            listOf(current) + nestedInstances
-          }
-      }
 }
